feat: add access permissions

+++ fastapi_app/src/domain/category/use_cases/delete_category.py
@@ -1,5 +1,6 @@
 from src.infrastructure.sqlite.database import database
 from src.infrastructure.sqlite.repositories.categories import CategoryRepository
+from src.core.exceptions.domain_exceptions import ForbiddenActionException
 
 
 class DeleteCategoryUseCase:
@@ -7,7 +8,9 @@ class DeleteCategoryUseCase:
         self._database = database
         self._repo = CategoryRepository()
 
-    async def execute(self, category_id: int) -> bool:
+    async def execute(self, category_id: int, is_superuser: bool = False) -> bool:
+        if not is_superuser:
+            raise ForbiddenActionException()
         with self._database.session() as session:
             self._repo.delete(session=session, id=category_id)
